+++ Sources/AriaLite/AriaLiteApp.swift
@@ -108,7 +108,7 @@ struct AriaLiteApp: App {
             SettingsWindowView()
                 .environmentObject(store)
                 .onAppear {
-                    AppPresentation.settingsDidAppear()
+                    AppPresentation.settingsDidAppear(store: store)
                 }
                 .onDisappear {
                     AppPresentation.settingsDidDisappear(store: store)